improvised to remove the syntax errors in the main, removed the payslip printing from the main and added it to the Payslip Controller to utilise the controller as it has not been used, payslip services now deemed unused so wil find a way to either utilise it or remove it, payslip controller modified to include the payslip, employee class -> payFreq now changed to int instead of double as it is a multiple choice value, added more advanced logic such as ?, !! and !, to protect the code from null values as nothing is predefined, !! was added to guarentee the program there will not be a null when it comes around, ! allows a statement to become null and ? lets the code know the vlaue may be null

+++ src/main/kotlin/models/Employee.kt
@@ -1,7 +1,7 @@
 package models
 
 data class Employee(val name: String,
-                    var payFreq: Double = 0.0,
+                    var payFreq: Int = 0,
                     var hourlyPay: Double = 0.0,
                     var sundayPay: Double = 0.0,
                     var taxPercentage: Double = 0.0
